refactor: extract useClassDays hook and fix styling precedence
- Extract shared useClassDays hook to reduce duplication
- Fix styling precedence so holidays take precedence over non-class-days
- A day that is both a holiday and non-class-day shows holiday styling
