+++ src/app/classrooms/[classroomId]/TeacherLessonCalendarTab.tsx
@@ -9,7 +9,8 @@ import { PageContent, PageLayout } from '@/components/PageLayout'
 import { getOntarioHolidays } from '@/lib/calendar'
 import { useRightSidebar } from '@/components/layout'
 import { lessonPlansToMarkdown, markdownToLessonPlans } from '@/lib/lesson-plan-markdown'
-import type { ClassDay, Classroom, LessonPlan, TiptapContent, Assignment } from '@/types'
+import { useClassDays } from '@/hooks/useClassDays'
+import type { Classroom, LessonPlan, TiptapContent, Assignment } from '@/types'
 import { writeCookie } from '@/lib/cookies'
 
 const AUTOSAVE_DEBOUNCE_MS = 3000
@@ -23,7 +24,7 @@ export function TeacherLessonCalendarTab({ classroom }: Props) {
   const router = useRouter()
   const [lessonPlans, setLessonPlans] = useState<LessonPlan[]>([])
   const [assignments, setAssignments] = useState<Assignment[]>([])
-  const [classDays, setClassDays] = useState<ClassDay[]>([])
+  const classDays = useClassDays(classroom.id)
   const [loading, setLoading] = useState(true)
   const [saving, setSaving] = useState(false)
   const [viewMode, setViewMode] = useState<CalendarViewMode>('week')
@@ -85,20 +86,6 @@ export function TeacherLessonCalendarTab({ classroom }: Props) {
       }
     }
     loadAssignments()
-  }, [classroom.id])
-
-  // Fetch class days for the classroom
-  useEffect(() => {
-    async function loadClassDays() {
-      try {
-        const res = await fetch(`/api/classrooms/${classroom.id}/class-days`)
-        const data = await res.json()
-        setClassDays(data.class_days || [])
-      } catch (err) {
-        console.error('Error loading class days:', err)
-      }
-    }
-    loadClassDays()
   }, [classroom.id])
 
   // Save a single lesson plan
